requirements.txt
discord.py>=1.7.3
aiohttp>=3.7.4
colorama>=0.4.4
